requirements.txt
python-telegram-bot==20.3
telethon==1.29.3
requests==2.31.0
